+++ src/core/crypto/index.ts
@@ -14,7 +14,6 @@ export * from "./privateKey";
 export * from "./publicKey";
 export * from "./secp256k1";
 export * from "./secp256r1";
-export * from "./slhDsaSha2128s";
 export * from "./signature";
 export * from "./singleKey";
 export * from "./types";