+++ tests/unit/hdKey.test.ts
@@ -1,12 +1,6 @@
 import { secp256k1WalletTestObject, wallet } from "./helper";
-import {
-  Ed25519PrivateKey,
-  Hex,
-  isValidBIP44Path,
-  isValidHardenedPath,
-  Secp256k1PrivateKey,
-  SlhDsaSha2128sPrivateKey,
-} from "../../src";
+import { Ed25519PrivateKey, Hex, isValidBIP44Path, isValidHardenedPath, Secp256k1PrivateKey } from "../../src";
+import { SlhDsaSha2128sPrivateKey } from "../../src/slh-dsa-sha2-128s";
 
 describe("Hierarchical Deterministic Key (hdkey)", () => {
   describe("hardened path", () => {